Add A2C configs.
We follow the steps taken by Huang et al. (2022) in implementing A2C as
a special case of PPO.

Because 'scale_reward' is set to False within the
existing configurations, we don't need to remove advantage
normalization, as it's already disabled. Moreover, entropy
regularization is not implemented in TRLX's implementation of PPO, so we
don't need to manually set the entropy coefficient to 0.