Support nested updates in `merge`
Currently `TRLConfig.update` will fail if given something like `{ "optimizer": { "name": "adam_8bit_bnb" } }" since it'll check if the `optimizer` key is in `updates`, but `merge` won't add it because the only keys updated are nested. Actually currently `merge` will check the values in updates in every nested dict, which is problematic when changing an attr that is common across many subconfigs (e.g `"name"`), so this way allows proper nested updates.